Move helper wallet reducers
