+++ packages/wallet/src/redux/initializedReducer.ts
@@ -5,14 +5,14 @@ import {
   INITIALIZING_CHANNEL,
   WAITING_FOR_CHANNEL_INITIALIZATION,
   channelInitialized,
-} from '../state';
+} from './state';
 
-import { WalletAction, CHANNEL_INITIALIZED } from '../actions';
+import { WalletAction, CHANNEL_INITIALIZED } from './actions';
 import { channelInitializationSuccess } from 'magmo-wallet-client/lib/wallet-events';
 import { ethers } from 'ethers';
-import { channelReducer } from './channels';
-import { unreachable } from '../../utils/reducer-utils';
-import { outboxStateReducer } from '../reducer';
+import { channelReducer } from './reducers/channels';
+import { unreachable } from '../utils/reducer-utils';
+import { outboxStateReducer } from './reducer';
 
 export const initializedReducer = (
   state: InitializedState,
